StateService emits data as it is received
