+++ src/app/shared/couchdb.service.ts
@@ -105,13 +105,20 @@ export class CouchService {
     }));
   }
 
-  findAll(db: string, query: any = { 'selector': { '_id': { '$gt': null } }, 'limit': 1000 }, opts?: any) {
-    console.log(query);
+  findAll(db: string, query: any, opts?: any) {
+    return this.findAllRequest(db, query, opts).pipe(flatMap(({ docs }) => docs), toArray());
+  }
+
+  findAllStream(db: string, query: any, opts?: any) {
+    return this.findAllRequest(db, query, opts).pipe(map(({ docs }) => docs));
+  }
+
+  private findAllRequest(db: string, query: any = { 'selector': { '_id': { '$gt': null } }, 'limit': 1000 }, opts?: any) {
     return this.post(db + '/_find', query, opts).pipe(expand((res) => {
       return this.post(db + '/_find', { ...query, bookmark: res.bookmark }, opts);
     }), takeWhile((res) => {
       return res.docs.length > 0;
-    }), flatMap(({ docs }) => docs), toArray());
+    }));
   }
 
   bulkGet(db: string, ids: string[], opts?: any) {
@@ -152,20 +159,6 @@ export class CouchService {
     local = parseInt(local.split('-')[0], 10);
     parent = parseInt(parent.split('-')[0], 10);
     return (local < parent) ? 'newerAvailable' : (local > parent) ? 'parentOlder' : 'mismatch';
-  }
-
-  combineChanges(docs: any[], changesDocs: any[]) {
-    return docs.reduce((newDocs: any[], doc: any) => {
-      const changesDoc = changesDocs.find((cDoc: any) => doc._id === cDoc._id);
-      if (changesDoc && changesDoc._deleted === true) {
-        return newDocs;
-      }
-      newDocs.push(changesDoc !== undefined ? changesDoc : doc);
-      return newDocs;
-    }, []).concat(
-      changesDocs.filter((cDoc: any) =>
-        cDoc._deleted !== true && docs.findIndex((doc: any) => doc._id === cDoc._id) === -1)
-    );
   }
 
   currentTime() {
